requirement.txt
beautifulsoup4==4.4.1
bs4==0.0.1
cycler==0.10.0
DateTime==4.1.1
EasyProcess==0.2.2
enum==0.4.6
lxml==3.6.0
matplotlib==1.5.1
nltk==3.2.1
numpy==1.11.1
Pillow==3.3.0
pymongo==2.8
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.6.1
PyVirtualDisplay==0.2
requests==2.10.0
scipy==0.17.1
selenium==2.53.6
six==1.10.0
textblob==0.11.1
wordcloud==1.2.1
xvfbwrapper==0.2.8
zope.interface==4.2.0
